Work on refactoring controller
