+++ controller/basic_scans.py
@@ -7,56 +7,7 @@ from datetime import datetime
 import nmap
 import sys
 
-# A list of all available tests. Update every time
-# a new test is created.
-def test_list():
-    tests = []
-    tests.append(create_json_test(1,"Nmap Port Scan - Common", "Common port scan", "Any"))
-    tests.append(create_json_test(2,"Nmap Port Scan - Full", "Full port scan", "Any"))
-    tests.append(create_json_test(3,"Eternal Blue Scan", "Searches for Eternal Blue vulnerability", "Windows 7, Windows 10"))
-    return tests
-
-def create_json_test(id, name, desc, os_list):
-    return {
-        "test_id": str(id),
-        "test_name": name,
-        "test_description": desc,
-        "operating_systems": os_list,
-    }
-
-# Creates the result object in JSON format
-def create_json_result(res_id, test_id, desc, date, val):
-    return {
-        "id": str(res_id),
-        "test_id": str(test_id),
-        "description": desc,
-        "date": date.strftime("%b %d %Y %H:%M:%S"),
-        "value" : val
-    }
-
 def eternal_blue_scan(test_id, host_ip, res_id):
-    date = datetime.now()
-    val = ""
-    desc = ""
-
-    scanner = nmap.PortScanner()
-    scanner.scan(hosts=host_ip, ports="445", arguments="--script smb-vuln-ms17-010")
-
-    if not scanner.all_hosts():
-        desc = "No hosts detected"
-        val = "Info"
-        return create_json_result(res_id, test_id, desc, date, val)
-
-    host = scanner.all_hosts()[0]
-    if "hostscript" not in scanner[host].keys():
-        desc = "Host is secure."
-        val = "Success"
-    else:
-        desc = """Host is vulnerable!
-===Nmap Output===
-""" + scanner[host]["hostscript"][0]["output"]
-        val = "Failure"
-    return create_json_result(res_id, test_id, desc, date, val)
 
 def basic_port_scan(test_id, host_ip, res_id):
     return port_scan(test_id, host_ip, res_id, "21-443")
@@ -96,24 +47,6 @@ def port_scan(test_id, host_ip, res_id, port_range):
     desc = str(open_ports)
     val = "Success"
     return create_json_result(res_id,test_id,desc,date,val)
-
-# Mapping of Test IDs to functions
-def run_test(id, ip, idx):
-    tests = {
-        1: basic_port_scan,
-        2: full_port_scan,
-        3: eternal_blue_scan,
-    }
-    return tests[id](id, ip, idx)
-
-# Loop through all of the Test IDs, run the corresponding tests, and return the results
-def execute_tests(ip, tests):
-    results = []
-    idx = 0
-    for id in tests:
-        results.append(run_test(int(id), ip, idx))
-        idx += 1
-    return results
 
 if __name__ == "__main__":
     args = sys.argv[1:] # Ignore the filename argument
